+++ src/app/(route)/feed/components/MainGoalItem/components/MainDDuDuInput/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useUpdateDDuDuMutation } from "./useUpdateDDuDuMutation/useUpdateDDuDuMutation";
